requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
streamlit>=1.28.0
python-binance>=1.0.16
pandas-ta>=0.3.14b
openpyxl>=3.0.9
websocket-client>=1.3.0
sqlite3
dataclasses